# remote_openapi.json
{"openapi":"3.1.0","info":{"title":"SDP Conventions API","description":"\nAPI REST pour acceder aux donnees des conventions collectives francaises.\n\nDeveloppe par **Smart Data Pay**.\n\n## Fonctionnalites\n- Liste de toutes les conventions collectives\n- Recherche par IDCC, nom, mot-cle\n- Acces au contenu detaille (sections, TOC, metadonnees)\n- Filtrage par section (ex: maladie, conges, salaire)\n- Synchronisation automatique des donnees\n","version":"1.0.0"},"paths":{"/conventions/":{"get":{"tags":["Conventions"],"summary":"Liste de toutes les conventions","description":"Retourne la liste resumee de toutes les conventions extraites.","operationId":"list_conventions_conventions__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":2000,"minimum":1,"description":"Nombre max de resultats","default":500,"title":"Limit"},"description":"Nombre max de resultats"},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"description":"Offset pour la pagination","default":0,"title":"Offset"},"description":"Offset pour la pagination"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/conventions/idcc/{idcc}":{"get":{"tags":["Conventions"],"summary":"Convention par IDCC","description":"Retourne une convention par son numero IDCC.","operationId":"get_by_idcc_conventions_idcc__idcc__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"idcc","in":"path","required":true,"schema":{"type":"string","title":"Idcc"}},{"name":"full","in":"query","required":false,"schema":{"type":"boolean","description":"Inclure le HTML brut des sections","default":false,"title":"Full"},"description":"Inclure le HTML brut des sections"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/conventions/file/{filename}":{"get":{"tags":["Conventions"],"summary":"Convention par nom de fichier","description":"Retourne une convention par son nom de fichier source.","operationId":"get_by_filename_conventions_file__filename__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/conventions/idcc/{idcc}/toc":{"get":{"tags":["Conventions"],"summary":"Sommaire d'une convention","description":"Retourne le sommaire (table des matieres) d'une convention.","operationId":"get_toc_conventions_idcc__idcc__toc_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"idcc","in":"path","required":true,"schema":{"type":"string","title":"Idcc"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/conventions/idcc/{idcc}/sections":{"get":{"tags":["Conventions"],"summary":"Sections d'une convention","description":"Retourne les sections d'une convention, avec filtre optionnel.","operationId":"get_sections_conventions_idcc__idcc__sections_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"idcc","in":"path","required":true,"schema":{"type":"string","title":"Idcc"}},{"name":"keyword","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filtrer les sections par mot-cle","title":"Keyword"},"description":"Filtrer les sections par mot-cle"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/conventions/idcc/{idcc}/metadata":{"get":{"tags":["Conventions"],"summary":"Metadonnees d'une convention","description":"Retourne uniquement les metadonnees d'une convention.","operationId":"get_metadata_conventions_idcc__idcc__metadata_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"idcc","in":"path","required":true,"schema":{"type":"string","title":"Idcc"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search/":{"get":{"tags":["Recherche"],"summary":"Recherche par nom ou IDCC","description":"Recherche dans les noms et numeros IDCC des conventions.\n\nExemples :\n- `/search/?q=metallurgie` → toutes les conventions metallurgie\n- `/search/?q=1486` → convention IDCC 1486\n- `/search/?q=hotel` → conventions hotellerie","operationId":"search_conventions_search__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","minLength":1,"description":"Terme de recherche (nom ou IDCC)","title":"Q"},"description":"Terme de recherche (nom ou IDCC)"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search/section":{"get":{"tags":["Recherche"],"summary":"Recherche dans les sections","description":"Recherche un mot-cle dans le contenu des sections de toutes les conventions.\n\nExemples :\n- `/search/section?keyword=maladie` → sections parlant de maladie\n- `/search/section?keyword=conges` → sections sur les conges\n- `/search/section?keyword=indemnite` → sections sur les indemnites","operationId":"search_in_sections_search_section_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string","minLength":2,"description":"Mot-cle a chercher dans les sections","title":"Keyword"},"description":"Mot-cle a chercher dans les sections"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["General"],"summary":"Root","description":"Page d'accueil de l'API.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}},"/stats":{"get":{"tags":["General"],"summary":"Statistiques","description":"Retourne les statistiques globales des donnees.","operationId":"get_stats_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}},"/sync":{"post":{"tags":["Administration"],"summary":"Synchroniser les donnees","description":"Synchronise les donnees depuis le dossier d'extraction vers le dossier de l'API.\nCopie les nouveaux fichiers et met a jour les fichiers modifies.","operationId":"sync_data_sync_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}},"/reload":{"post":{"tags":["Administration"],"summary":"Recharger les donnees","description":"Recharge les donnees depuis les fichiers locaux (sans synchronisation).","operationId":"reload_data_reload_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}},"/list":{"get":{"tags":["General"],"summary":"Liste brute des conventions","description":"Retourne la liste brute des conventions (depuis conventions_list.json).","operationId":"get_conventions_list_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-API-Key"}}}}